fix: Resolve unrefreshable custom twibbon /go page
